KOT-X-Backend/requirements.txt
asgiref==3.11.1
asttokens==3.0.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
backcall==0.2.0
beautifulsoup4==4.14.3
black==26.3.1
bleach==6.3.0
cbor2==5.8.0
certifi==2026.2.25
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.5
daphne==4.2.1
defusedxml==0.7.1
distlib==0.4.0
Django==5.2.3
django-cors-headers==4.9.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
docopt==0.6.2
executing==2.2.1
Faker==40.11.0
faker_food==0.3.0
fastjsonschema==2.21.2
filelock==3.25.2
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
ipython==8.12.3
isort==8.0.1
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.2.0
msgpack==1.1.2
mypy_extensions==1.1.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
packaging==26.0
pandocfilters==1.5.1
parso==0.8.6
pathspec==1.0.4
pickleshare==0.7.5
pillow==12.1.1
platformdirs==4.9.4
prompt_toolkit==3.0.52
pure_eval==0.2.3
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
Pygments==2.19.2
PyJWT==2.11.0
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-discovery==1.2.1
pytokens==0.4.1
pyzmq==27.1.0
redis==7.3.0
referencing==0.37.0
requests==2.33.1
rich==14.3.3
rpds-py==0.30.0
service-identity==24.2.0
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
stack-data==0.6.3
tinycss2==1.4.0
tornado==6.5.5
traitlets==5.14.3
Twisted==23.10.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
urllib3==2.6.3
virtualenv==21.2.0
wcwidth==0.6.0
webencodings==0.5.1
whitenoise==6.12.0
yarg==0.1.9
zope.interface==8.2
